% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{countToTpm}
\alias{countToTpm}
\title{Function from https://haroldpimentel.wordpress.com/2014/05/08/what-the-fpkm-a-review-rna-seq-expression-units/}
\usage{
countToTpm(counts, effLen)
}
\description{
Calculating TPM for each sample.
}
