% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_subset}
\alias{create_subset}
\title{PIVOT Data Control}
\usage{
create_subset(
  r_data,
  input,
  flist,
  slist,
  keep_filter = F,
  renorm = F,
  erccStds = NULL,
  actionType = c("Filter", "Subset"),
  actionText = ""
)
}
\description{
Create subset by filtering or subsetting
}
