% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_function.R
\name{custom_downloadbtn}
\alias{custom_downloadbtn}
\title{PIVOT Data Input, UI function}
\usage{
custom_downloadbtn(outputId, label = "Download", class = NULL,
  style = NULL)
}
\description{
PIVOT Data Input, UI function
}
