% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_session.R
\name{detachAllPackages}
\alias{detachAllPackages}
\title{Function used for clean the current R session}
\source{
\url{http://r.789695.n4.nabble.com/Is-it-possible-to-increase-MAX-NUM-DLLS-in-future-R-releases-td4720352.html}
}
\usage{
detachAllPackages()
}
\description{
This function calls unloadNamespaces by Henrik Bengtsson to unload all attached namespaces except the base ones.
}
\examples{
detachAllPackages()
}
