% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_function.R
\name{dropdownMenu}
\alias{dropdownMenu}
\title{PIVOT Data Input, UI function}
\usage{
dropdownMenu(
  ...,
  type = c("messages", "notifications", "tasks", "system", "support"),
  top_msg = NULL,
  icon = NULL,
  .list = NULL
)
}
\description{
PIVOT Data Input, UI function
}
