% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_function.R
\name{enhanced_box}
\alias{enhanced_box}
\title{PIVOT Data Input, UI function}
\usage{
enhanced_box(..., title = NULL, id = NULL, footer = NULL,
  status = NULL, solidHeader = FALSE, background = NULL, width = 6,
  height = NULL, collapsible = FALSE, reportable = FALSE,
  get_pdf = FALSE, get_html = FALSE, register_analysis = FALSE,
  collapsed = FALSE)
}
\description{
PIVOT Data Input, UI function
}
