% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{estimateSizeFactorsForMatrix_groupwise}
\alias{estimateSizeFactorsForMatrix_groupwise}
\title{Groupwise modified DESeq script}
\usage{
estimateSizeFactorsForMatrix_groupwise(counts, group, locfunc = median,
  geoMeans, controlGenes, threshold = 0.7)
}
\description{
Groupwise modified DESeq script
}

