% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{estimateSizeFactorsForMatrix_MK}
\alias{estimateSizeFactorsForMatrix_MK}
\title{Mugdhak's modified DESeq script}
\usage{
estimateSizeFactorsForMatrix_MK(counts, locfunc = median, geoMeans,
  controlGenes, threshold = 0.7)
}
\description{
Mugdhak's modified DESeq script
}

