% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot.R
\name{feature_super_plot}
\alias{feature_super_plot}
\title{Plot feature expression with ggplot}
\usage{
feature_super_plot(df, selected_gene, plot_by = "sample", meta = NULL,
  palette = "Set1", style = "box", log_scale = F, legend_pos = "top",
  textSize = 15, pointSize = 3)
}
\description{
Plot feature expression with ggplot
}
