% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_map.R
\name{generate_data_node}
\alias{generate_data_node}
\title{Generate data history node and edge table for data map}
\usage{
generate_data_node(his_tbl = NULL, reg_tbl = NULL)
}
\description{
Generate data history node and edge table for data map
}
