% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{generate_mst}
\alias{generate_mst}
\title{Gnerate minimum spanning tree given distance matrix}
\usage{
generate_mst(dist_mtx, method = "walktrap", step = NULL,
  color_list = NULL)
}
\description{
Gnerate minimum spanning tree given distance matrix
}
