% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scale.R
\name{generateRelativeFreq}
\alias{generateRelativeFreq}
\title{PIVOT generate relative frequency table given counts data}
\usage{
generateRelativeFreq(raw)
}
\description{
PIVOT generate relative frequency table given counts data
}

