% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot.R
\name{get_brewer_set}
\alias{get_brewer_set}
\title{Generate color set for plots}
\usage{
get_brewer_set(palette = c("sequential", "diverging", "qualitative"))
}
\description{
Generate color set for plots
}

