% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot.R
\name{get_color_vector}
\alias{get_color_vector}
\title{Get a vector of colors with RColorBrewer for group labels}
\usage{
get_color_vector(labels, pal = "Set1", maxCol = 9)
}
\description{
Get a vector of colors with RColorBrewer for group labels
}

