% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_function.R
\name{ggplotly}
\alias{ggplotly}
\title{ggplotly from plotly}
\usage{
ggplotly(p = ggplot2::last_plot(), width = NULL, height = NULL,
  tooltip = "all", dynamicTicks = FALSE, layerData = 1,
  originalData = TRUE, source = "A", ...)
}
\description{
ggplotly from plotly
}

