% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_function.R
\name{infoBoxOutput_custom}
\alias{infoBoxOutput_custom}
\title{PIVOT Data Input, UI function}
\usage{
infoBoxOutput_custom(outputId, width = 4, style = NULL)
}
\description{
PIVOT Data Input, UI function
}

