% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_function.R
\name{makeCheckboxTooltip}
\alias{makeCheckboxTooltip}
\title{Checkbox tooltip, from K. Rohde
Source: http://stackoverflow.com/questions/36670065/tooltip-in-shiny-ui-for-help-text}
\usage{
makeCheckboxTooltip(checkboxValue, buttonLabel, Tooltip)
}
\description{
Checkbox tooltip, from K. Rohde
Source: http://stackoverflow.com/questions/36670065/tooltip-in-shiny-ui-for-help-text
}
