% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plot.R
\name{meanSdPlot}
\alias{meanSdPlot}
\title{meanSDplot}
\usage{
meanSdPlot(x, ranks = TRUE, xlab = ifelse(ranks, "rank(mean)", "mean"),
  ylab = "sd", pch, plot = TRUE, bins = 50, ...)
}
\description{
meanSDplot function from VSN package
}
