% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_function.R
\name{modalActionButton}
\alias{modalActionButton}
\title{PIVOT Data Input, UI function}
\usage{
modalActionButton(inputId, label, icon = NULL, width = NULL, ...)
}
\description{
custom modal button which also serves as an action button
}
