% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_function.R
\name{modalConfirmDialog}
\alias{modalConfirmDialog}
\title{PIVOT Data Input, UI function}
\usage{
modalConfirmDialog(id, header = "Confirmation", body = "Are you sure?",
  footer = list(actionButton("confirmDlgOkBtn", "OK")))
}
\description{
Modal confirm button from Weicheng Zhu
source: https://github.com/mingsnu/shiny-confirm-dialog
Note: This was not used in PIVOT. I later switched to shiny modal.
}

