% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_function.R
\name{modalTriggerButton}
\alias{modalTriggerButton}
\title{PIVOT Data Input, UI function}
\usage{
modalTriggerButton(inputId, target, label, icon = NULL,
  class = "btn action-button btn-danger btn_leftAlign")
}
\description{
modal trigger button from Weicheng Zhu
source: https://github.com/mingsnu/shiny-confirm-dialog
Note: This was not used in PIVOT. I later switched to shiny modal.
}
