% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_function.R
\name{numericInput_1}
\alias{numericInput_1}
\title{PIVOT Data Input, UI function
custom numeric input based on shiny}
\usage{
numericInput_1(inputId, label, value, min = NA, max = NA, step = NA,
  width = NULL)
}
\description{
PIVOT Data Input, UI function
custom numeric input based on shiny
}

