% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_corr.R
\name{pairs_extended}
\alias{pairs_extended}
\title{Function for generating pairwise correlation plot, from SCDE}
\usage{
pairs_extended(
  x,
  labels,
  panel = points,
  ...,
  lower.panel = panel,
  upper.panel = panel,
  diag.panel = NULL,
  text.panel = textPanel,
  label.pos = 0.5 + has.diag/3,
  cex.labels = NULL,
  font.labels = 1,
  row1attop = TRUE,
  gap = 1
)
}
\description{
Function for generating pairwise correlation plot, from SCDE
}
