% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{pivot_analysis}
\alias{pivot_analysis}
\title{Launch PIVOT Data Basic Analysis Module}
\usage{
pivot_analysis(r_module = NULL)
}
\description{
Launch PIVOT Data Basic Analysis Module
}
\examples{
pivot_analysis()
}
