% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot.R
\name{pivot_colorBy}
\alias{pivot_colorBy}
\title{Choose which meta column to be used as coloring group}
\usage{
pivot_colorBy(input, output, session, meta)
}
\description{
This is the server part of the module
}
