% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scale.R
\name{pivot_dataScale}
\alias{pivot_dataScale}
\title{PIVOT data scale module}
\usage{
pivot_dataScale(input, output, session, r_data, order_by = NULL,
  keep_stats = FALSE)
}
\description{
Controls which transformed data should be used as input for analysis modules, this is the server part of the module.
}

