% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scale.R
\name{pivot_dataScale_UI}
\alias{pivot_dataScale_UI}
\title{PIVOT data scale module}
\usage{
pivot_dataScale_UI(id, include = c("Counts (raw)", "Counts (normalized)",
  "Relative Frequency", "Log10 Counts", "Standardized Counts",
  "Log10 & Standardized"), selected = "Counts (normalized)")
}
\description{
Controls which transformed data should be used as input for analysis modules, this is the UI part of the module.
}
