% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scale.R
\name{pivot_dataScaleRange}
\alias{pivot_dataScaleRange}
\title{a wrapper module for scaling and range selection for heatmap}
\usage{
pivot_dataScaleRange(input, output, session, r_data, keep_stats = FALSE)
}
\description{
This is the server part of the module.
}

