% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_group.R
\name{pivot_deGroupBy}
\alias{pivot_deGroupBy}
\title{Choose the design model for DE analysis}
\usage{
pivot_deGroupBy(input, output, session, meta, reduced = c("no", "yes",
  "maybe"))
}
\description{
This is the server part of the module
}

