% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_plot.R
\name{pivot_dimScatter}
\alias{pivot_dimScatter}
\title{a wrapper module for scatter plot in 1D, 2D and 3D with different packages}
\usage{
pivot_dimScatter(input, output, session, type = c("pca", "tsne", "plda"),
  obj = NULL, minfo = NULL)
}
\description{
This is the server part of the module.
}
