% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_plot.R
\name{pivot_dimScatter_UI}
\alias{pivot_dimScatter_UI}
\title{a wrapper module for plot in 1D, 2D and 3D with different packages}
\usage{
pivot_dimScatter_UI(id, type = c("pca", "tsne", "plda"))
}
\description{
This is the UI part of the module.
}
