% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{pivot_featureInputModal}
\alias{pivot_featureInputModal}
\title{a wrapper for feature input and preview in Modal}
\usage{
pivot_featureInputModal(input, output, session, r_data, match_rdata = T)
}
\description{
This is the server part of the module.
}

