% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{pivot_featureInputModal_UI}
\alias{pivot_featureInputModal_UI}
\title{a wrapper for feature input and preview in Modal}
\usage{
pivot_featureInputModal_UI(id, label = NULL)
}
\description{
This is the ui part of the module.
}
