% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scale.R
\name{pivot_featureList}
\alias{pivot_featureList}
\title{shiny module for choosing feature set generated by DE or custom input}
\usage{
pivot_featureList(input, output, session, r_data)
}
\description{
This is the server part of the module.
}

