% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scale.R
\name{pivot_featureList_UI}
\alias{pivot_featureList_UI}
\title{shiny module for choosing feature set generated by DE or custom input}
\usage{
pivot_featureList_UI(
  id,
  include = c("custom", "deseq", "edgeR", "scde", "monocle", "mwu"),
  selected = "custom",
  width = 8
)
}
\description{
This is the ui part of the module.
}
