% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot.R
\name{pivot_featurePlot}
\alias{pivot_featurePlot}
\title{The feature plot module of PIVOT}
\usage{
pivot_featurePlot(input, output, session, meta, df, gene, ids = NULL)
}
\description{
This is the server part of the module
}

