% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot.R
\name{pivot_featurePlot_UI}
\alias{pivot_featurePlot_UI}
\title{The feature plot module of PIVOT}
\usage{
pivot_featurePlot_UI(id, meta, ids = NULL)
}
\description{
This is the UI part of the module, includes scale option, color group option, color set option, plot style option and the plot.
}
