% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{pivot_filePreview}
\alias{pivot_filePreview}
\title{PIVOT file input preview module, server}
\usage{
pivot_filePreview(input, output, session, df, height = "350px", search = FALSE)
}
\description{
PIVOT file input preview module, server
}
