% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot.R
\name{pivot_groupBy_UI}
\alias{pivot_groupBy_UI}
\title{Choose which meta column to be used as coloring group}
\usage{
pivot_groupBy_UI(id, category, append_sample = T, append_none = F,
  multiple = F, choose_color = T, bset = "qualitative", width = 8)
}
\description{
This is the UI part of the module
}

