% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher.R
\name{pivot_launcher}
\alias{pivot_launcher}
\title{PIVOT laucher}
\usage{
pivot_launcher()
}
\description{
Serves as a gateway for all analysis modules and allow users to monitor the R session state and clean when needed.
}
\examples{
pivot_launcher()
}
