% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_plot.R
\name{pivot_Plot2d}
\alias{pivot_Plot2d}
\title{a wrapper module for plot in 2D}
\usage{
pivot_Plot2d(input, output, session, type = NULL, obj = NULL, proj = NULL,
  minfo = NULL, source = NULL, event = NULL, selected = NULL)
}
\description{
This is the server part of the module.
}
