% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_plot.R
\name{pivot_Plot3d}
\alias{pivot_Plot3d}
\title{a wrapper module for plot in 3D}
\usage{
pivot_Plot3d(input, output, session, type = NULL, obj = NULL, proj = NULL,
  minfo = NULL)
}
\description{
This is the server part of the module.
}

