% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_function.R
\name{pivotboardHeader}
\alias{pivotboardHeader}
\title{PIVOT Data Input, UI function from shinydashboard
Modified for sidebar-mini}
\usage{
pivotboardHeader(
  ...,
  title = NULL,
  titleWidth = NULL,
  disable = FALSE,
  .list = NULL
)
}
\description{
PIVOT Data Input, UI function from shinydashboard
Modified for sidebar-mini
}
