% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ercc_plot.R
\name{plotERCCDistribution}
\alias{plotERCCDistribution}
\title{Plot ERCC distribution and perform linear regression}
\usage{
plotERCCDistribution(counts)
}
\description{
Plot ERCC distribution and perform linear regression
}
