% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_corr.R
\name{scde_pair_plot}
\alias{scde_pair_plot}
\title{Function for generating pairwise correlation plot, from SCDE}
\usage{
scde_pair_plot(counts, groups, rl, ids)
}
\description{
Function for generating pairwise correlation plot, from SCDE
}
