% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_map.R
\name{search_subset_node}
\alias{search_subset_node}
\title{PIVOT data management function}
\usage{
search_subset_node(tbl, node)
}
\description{
This function is for determining which sample subset the filter operation is performed on, if not keep filtering.
}

