% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot.R
\name{sql_unpivot}
\alias{sql_unpivot}
\title{Create an unpivot query}
\usage{
sql_unpivot(con, from, select, key, value, levels, order_by = NULL)
}
\arguments{
\item{con}{a Database connection}

\item{from}{the from clause}

\item{select}{variables to select in addition to levels.}

\item{key}{Variable columns originate from}

\item{value}{The expression to evaluate to create the values}

\item{levels}{the columns to turn into values of a variable.}

\item{order_by}{optional order by clause}
}
\description{
Creates a SQL pivot query.  Similar to the \code{tidyr::gather}
function.
}
