% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_function.R
\name{systemItem}
\alias{systemItem}
\title{PIVOT Data Input, UI function}
\usage{
systemItem(title = NULL, text, icon = NULL, status = "success",
  href = NULL)
}
\description{
PIVOT Data Input, UI function
}
