% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot.R
\name{unpivot_query}
\alias{unpivot_query}
\title{Create a pivot query representation}
\usage{
unpivot_query(from, key, value, levels, select = character(0),
  order_by = NULL)
}
\arguments{
\item{from}{the from clause}

\item{key}{Variable columns originate from}

\item{value}{The expression to evaluate to create the values}

\item{levels}{the columns to turn into values of a variable.}

\item{select}{variables to select in addition to levels.}

\item{order_by}{optional order by clause}
}
\description{
Create a pivot table
}
