% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_map.R
\name{update_history}
\alias{update_history}
\title{PIVOT data management function}
\usage{
update_history(
  r_data,
  parent,
  action_type,
  action,
  lists,
  norm_method,
  norm_param
)
}
\description{
Write new entries to data history
}
