% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_function.R
\name{valueBoxOutput_custom}
\alias{valueBoxOutput_custom}
\title{PIVOT Data Input, UI function}
\usage{
valueBoxOutput_custom(outputId, width = 4, style = NULL)
}
\description{
PIVOT Data Input, UI function
}
