% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_group.R
\name{pivot_deGroupBy_UI}
\alias{pivot_deGroupBy_UI}
\title{Choose the design model for DE analysis}
\usage{
pivot_deGroupBy_UI(id, meta, width = 12, reduced = c("no", "yes",
  "maybe"), model = c("condition", "condition_batch", "custom"))
}
\description{
This module allow user choose which model, e.g., "~ condition", "~condition + batch"  or "timecourse", should be used for DE analysis,
and which meta data column should be used as the covariates.
}
