% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{sql_pivot}
\alias{sql_pivot}
\title{Create a Pivot Query}
\usage{
sql_pivot(con, from, select, key, value, levels, ...)
}
\arguments{
\item{con}{a Database connection}

\item{from}{the from clause}

\item{select}{variables to select in addition to levels.}

\item{key}{Variable columns originate from}

\item{value}{The expression to evaluate to create the values}

\item{levels}{the levels of key to turn into columns.}

\item{...}{arguments to pass on or ignore.}
}
\description{
Creates a SQL pivot query.  Similar to the \code{tidyr::spread}
function.
}
