% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot.R
\name{unpivot}
\alias{unpivot}
\title{Pivot a table}
\usage{
unpivot(data, key, value, ...)
}
\arguments{
\item{data}{A data frame.}

\item{key}{Column names or positions. This is passed to
\code{\link[tidyselect:vars_pull]{tidyselect::vars_pull()}}.

These arguments are passed by expression and support
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names or column positions).}

\item{value}{Column names or positions. This is passed to
\code{\link[tidyselect:vars_pull]{tidyselect::vars_pull()}}.

These arguments are passed by expression and support
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names or column positions).}

\item{...}{Selection criteria for levels of key to select.}
}
\description{
Pivot a table
}
\examples{
\dontrun{
# establish db as a database connection
library(dbplyr)
library(tidyverse)
db_iris <- copy_to(db, iris)
long.iris <- unpivot(db_iris, Variable, Value, Sepal.Length, Sepal.Width, Petal.Length, Petal.Width)
means <- summarise(group_by(long.iris, Species, Variable), Mean = mean(Value, na.rm=TRUE)
pivot(means, Species, setosa, versicolor, virginica)
}

}
