% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot.R
\name{unpivot}
\alias{unpivot}
\title{Pivot a table}
\usage{
unpivot(data, key, value, ...)
}
\arguments{
\item{data}{A data frame.}

\item{key}{Column names or positions. This is passed to
\code{\link[tidyselect:vars_pull]{tidyselect::vars_pull()}}.

These arguments are passed by expression and support
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names or column positions).}

\item{value}{Column names or positions. This is passed to
\code{\link[tidyselect:vars_pull]{tidyselect::vars_pull()}}.

These arguments are passed by expression and support
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names or column positions).}

\item{...}{Selection criteria for levels of key to select.}
}
\description{
Pivot a table
}
